Move build steps into build system
