+++ src/pipelines/html.rs
@@ -5,14 +5,12 @@ use std::sync::Arc;
 
 use anyhow::{anyhow, ensure, Context, Result};
 use async_std::fs;
-use async_std::path::Path;
 use async_std::task::{spawn_local, JoinHandle};
 use futures::channel::mpsc::Sender;
 use futures::stream::{FuturesUnordered, StreamExt};
 use indicatif::ProgressBar;
 use nipper::Document;
 
-use crate::common::{copy_dir_recursive, remove_dir_all};
 use crate::config::RtcBuild;
 use crate::pipelines::rust_app::RustApp;
 use crate::pipelines::{TrunkLink, TrunkLinkPipelineOutput, TRUNK_ID};
@@ -65,8 +63,6 @@ impl HtmlPipeline {
 
     /// Perform the build routine of this pipeline.
     async fn build(self: Arc<Self>) -> Result<()> {
-        self.prepare_staging_dist().await.context("error preparing build environment")?;
-
         self.progress.clone().set_message("spawning asset pipelines");
 
         // Open the source HTML file for processing.
@@ -117,42 +113,6 @@ impl HtmlPipeline {
             .await
             .context("error writing finalized HTML output")?;
 
-        self.apply_dist().await.context("error applying built distribution")?;
-        Ok(())
-    }
-
-    /// Moves the contents of dist/.current into dist, signifying the application
-    /// of a successful build. Also removes dist/.current afterwards.
-    async fn apply_dist(self: Arc<Self>) -> Result<()> {
-        let final_dist = self.cfg.final_dist.clone();
-        let staging_dist = self.cfg.staging_dist.clone();
-        self.progress.clone().set_message("applying new distribution");
-
-        // build succeeded, so delete everything in `dist`,
-        // copy everything from `dist/.current` to `dist`, and
-        // then delete `dist/.current`
-        let mut entries = fs::read_dir(&final_dist).await.context("error reading dist dir")?;
-        while let Some(entry) = entries.next().await {
-            let entry = entry.context("error reading contents of dist dir")?;
-            if entry.file_name() == ".current" {
-                continue;
-            }
-
-            let file_type = entry.file_type().await.context("error reading metadata of file in dist dir")?;
-
-            if file_type.is_dir() {
-                remove_dir_all(entry.path().into()).await.context("error cleaning dist")?;
-            } else if file_type.is_symlink() || file_type.is_file() {
-                fs::remove_file(entry.path()).await.context("error cleaning dist")?;
-            }
-        }
-
-        copy_dir_recursive(staging_dist.to_path_buf(), final_dist.to_path_buf())
-            .await
-            .context("error copying dist/.current dir to dist dir")?;
-
-        remove_dir_all(staging_dist).await.context("error deleting dist/.current")?;
-
         Ok(())
     }
 
@@ -171,30 +131,5 @@ impl HtmlPipeline {
         let mut base_elements = target_html.select(&format!("html head base[{}]", PUBLIC_URL_MARKER_ATTR));
         base_elements.remove_attr(PUBLIC_URL_MARKER_ATTR);
         base_elements.set_attr("href", &self.cfg.public_url);
-    }
-
-    /// Creates a "holding area" (dist/.current) for storing intermediate build results
-    async fn prepare_staging_dist(&self) -> Result<()> {
-        // Prepare holding area in which we will assemble the latest build
-        let staging_dist: &Path = self.cfg.staging_dist.as_path().into();
-
-        if staging_dist.exists().await {
-            // Clean holding area, if applicable
-            let mut entries = fs::read_dir(staging_dist).await.context("error reading dist/.current dir")?;
-            while let Some(entry) = entries.next().await {
-                let entry = entry.context("error reading contents of dist/.current dir")?;
-                let file_type = entry.file_type().await.context("error reading metadata of file in dist/.current dir")?;
-
-                if file_type.is_dir() {
-                    fs::remove_dir_all(entry.path()).await.context("Cleaning dist/.current failed")?;
-                } else if file_type.is_symlink() || file_type.is_file() {
-                    fs::remove_file(entry.path()).await.context("Cleaning dist/.current failed")?;
-                }
-            }
-        } else {
-            fs::create_dir_all(staging_dist).await.context("error creating dist/.current dir")?;
-        }
-
-        Ok(())
     }
 }
